add logger in chat-server

+++ chat-server/internal/api/delete.go
@@ -2,7 +2,6 @@ package api
 
 import (
 	"context"
-	"log"
 
 	desc "github.com/GolZrd/micro-chat/chat-server/pkg/chat_v1"
 
@@ -12,13 +11,14 @@ import (
 )
 
 func (s *Implementation) Delete(ctx context.Context, req *desc.DeleteRequest) (*emptypb.Empty, error) {
+	if req.Id == 0 {
+		return nil, status.Error(codes.InvalidArgument, "chat id is required")
+	}
 
 	err := s.chatService.Delete(ctx, req.Id)
 	if err != nil {
 		return nil, status.Errorf(codes.Internal, "failed to delete chat: %v", err)
 	}
-
-	log.Printf("Delete chat with id: %d", req.Id)
 
 	return &emptypb.Empty{}, nil
 }
